MDEV-5492 - Reduce usage of LOCK_open: TABLE::in_use
- moved TABLE::in_use updates out of TDC_element::LOCK_table_share
- use atomic load/store so that foreign threads accessing TABLE::in_use get
  consistent value
- relaxed assertions so that TABLE::in_use == 0 is now acceptable and is skipped
  by kill_delayed_threads_for_table() and TABLE_SHARE::visit_subgraph()

In all cases it is guaranteed that THD object stays valid until
TDC_element::LOCK_table_share is locked or TDC_element::all_tables_refs is
not 0.

It is safe to skip TABLE::in_use == 0 in kill_delayed_threads_for_table() and
TABLE_SHARE::visit_subgraph() because TABLE is either being released or it is
being acquired and will be released immediately.